// src/app/models/resUser.ts
export interface ResUser {
    cedula:           number;
    apellido1:        string;
    apellido2:        string;
    nombre1:          string;
    nombre2:          string;
    email:            string;
    password:         number;
    telefono:         string;
    direccion:        string;
    sexo:             string;
    fnacimiento:      Date;
    firma:            string;
    rol:              string;
    idAdministrativo: number;
}